+++ CHANGES.md
@@ -4,7 +4,7 @@ Change log for the StorPool OpenStack integration
 1.1.0.dev1
 ----------
 
-- Add the Ocata Cinder volume driver.
+- Add the Ocata Cinder volume driver and Nova attachment driver.
 
 1.0.0
 -----